+++ CHANGELOG.md
@@ -26,6 +26,8 @@ Other blocks are:
 
 ## Unreleased
 
+### Added
+- **VIZ**: `Molecules.animation_status` provides a getter/setter on the animation status (running, paused, stopped...).
 
 ## [0.6.0] - 2025-02-16